--- src/components/Tab/Tab.tsx
@@ -1,5 +1,0 @@
-const Tab = () => {
-  return <div></div>
-}
-
-export default Tab
