+++ src/components/common/Tab/Tab.tsx
@@ -1,0 +1,9 @@
+interface TabProps {
+  children?: React.ReactNode
+}
+
+const Tab = ({ children }: TabProps) => {
+  return <div className="flex transition ease-in-out">{children}</div>
+}
+
+export default Tab
